feat(consts): move constants in shared
