+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "@types/react": "^17.0.30",
+    "@types/react-router": "^5.1.20",
     "ajv": "^8.9.0",
     "axios": "^0.24.0",
     "chart.js": "^3.5.1",
@@ -51,6 +52,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^18.0.11",
     "gh-pages": "^3.2.3",
     "react-router-dom": "^6.2.1"
   }
