--- src/App.css
@@ -1,5 +1,0 @@
-
-h2 {
-    font-family: 'Roboto', sans-serif;
-}
-
